pyproject.toml
[tool.poetry]
name = "thrivey-api"
version = "0.1.0"
description = "A pydantic take on FastAPI"
authors = ["tim frazer <tim@littlebigthings.co>", "sean glynn <seanglynn06@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.8.1"
datamodel-code-generator = "^0.8.3"
bpython = "^0.21"
uvicorn = "^0.13.4"
SQLAlchemy = "^1.4.26"
Faker = "^9.8.0"
pandas = "^1.3.4"
databases = "^0.5.3"
aiosqlite = "^0.17.0"
python-decouple = "^3.5"

[tool.poetry.dev-dependencies]
typer = "^0.3.2"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
